Moved back weight_sources in classical
