+++ openquake/calculators/classical.py
@@ -16,6 +16,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
 import io
+import time
 import copy
 import psutil
 import pprint
@@ -31,7 +32,11 @@ from openquake.baselib.python3compat import encode
 from openquake.baselib.general import (
     AccumDict, DictArray, block_splitter, groupby, humansize,
     get_nbytes_msg)
+from openquake.hazardlib.source.point import (
+    PointSource, grid_point_sources, msr_name)
+from openquake.hazardlib.sourceconverter import SourceGroup
 from openquake.hazardlib.contexts import ContextMaker, get_effect
+from openquake.hazardlib.calc.filters import split_source
 from openquake.hazardlib.calc.hazard_curve import classical as hazclassical
 from openquake.hazardlib.probability_map import ProbabilityMap
 from openquake.commonlib import calc, util, readinput
@@ -43,6 +48,7 @@ U32 = numpy.uint32
 F32 = numpy.float32
 F64 = numpy.float64
 TWO32 = 2 ** 32
+EPS = .01
 get_weight = operator.attrgetter('weight')
 grp_extreme_dt = numpy.dtype([('et_id', U16), ('grp_trt', hdf5.vstr),
                              ('extreme_poe', F32)])
@@ -62,14 +68,62 @@ def get_extreme_poe(array, imtls):
     return max(array[imtls(imt).stop - 1].max() for imt in imtls)
 
 
-#  ########################### task functions ############################ #
+def get_factor(n, location, sitecol, psdist, maxdist):
+    """
+    :param n: an integer > 1 related to the number of ruptures in a source
+    :param location: the location of a source
+    :param sitecol: the SiteCollection
+    :param psdist: point source distance
+    :param maxdist: the maximum distance
+    :returns:
+        reduction factor in the range 1/n (if all sites are far away) to 1
+        (if all sites are close to the location)
+    """
+    close, far = sitecol.count_close_far(location, psdist, maxdist)
+    factor = (close + (far + EPS) / n) / (close + far + EPS)
+    return factor
 
-def classical(srcs, rlzs_by_gsim, params, monitor):
-    """
-    Read the SourceFilter and call the classical calculator in hazardlib
-    """
-    srcfilter = monitor.read('srcfilter')
-    return hazclassical(srcs, srcfilter, rlzs_by_gsim, params, monitor)
+
+def _weight_sources(csm, oqparam, h5):
+
+    # do nothing for atomic sources except counting the ruptures
+    for src in csm.get_sources(atomic=True):
+        src.num_ruptures = src.count_ruptures()
+        src.nsites = len(csm.srcfilter.sitecol)
+
+    # run weight_sources for non-atomic sources
+    sources_by_grp = groupby(
+        csm.get_sources(atomic=False),
+        lambda src: (src.grp_id, msr_name(src)))
+    param = dict(maximum_distance=oqparam.maximum_distance,
+                 pointsource_distance=oqparam.pointsource_distance,
+                 ps_grid_spacing=oqparam.ps_grid_spacing,
+                 split_sources=oqparam.split_sources)
+
+    res = parallel.Starmap(
+        weight_sources,
+        ((srcs, csm.srcfilter, param) for srcs in sources_by_grp.values()),
+        h5=h5, distribute=None if len(sources_by_grp) > 1 else 'no').reduce()
+
+    if res and res['before'] != res['after']:
+        logging.info('Reduced the number of sources from {:_d} -> {:_d}'.
+                     format(res['before'], res['after']))
+
+    if res and h5:
+        csm.update_source_info(res['calc_times'], nsites=True)
+
+    for grp_id, srcs in res.items():
+        # srcs can be empty if the minimum_magnitude filter is on
+        if srcs and not isinstance(grp_id, str):
+            newsg = SourceGroup(srcs[0].tectonic_region_type)
+            newsg.sources = srcs
+            csm.src_groups[grp_id] = newsg
+
+    # sanity check
+    for sg in csm.src_groups:
+        for src in sg:
+            assert src.num_ruptures
+            assert src.nsites
 
 
 def store_ctxs(dstore, rupdata, grp_id):
@@ -89,6 +143,59 @@ def store_ctxs(dstore, rupdata, grp_id):
                 dstore[n].resize((len(dstore[n]) + nr,))
         else:
             hdf5.extend(dstore[n], rupdata.get(par, nans))
+
+
+#  ########################### task functions ############################ #
+
+def weight_sources(srcs, srcfilter, params, monitor):
+    """
+    Weight the sources. Also split them if split_sources is true.
+    """
+    # nrups, nsites, time, task_no
+    calc_times = AccumDict(accum=numpy.zeros(4, F32))
+    sources = []
+    grp_id = srcs[0].grp_id
+    trt = srcs[0].tectonic_region_type
+    md = params['maximum_distance'](trt)
+    pd = (params['pointsource_distance'](trt)
+          if params['pointsource_distance'] else 0)
+    for src in srcs:
+        t0 = time.time()
+        trt = src.tectonic_region_type
+        splits = split_source(src) if params['split_sources'] else [src]
+        sources.extend(splits)
+        nrups = src.count_ruptures()
+        dt = time.time() - t0
+        calc_times[src.id] += F32([nrups, src.nsites, dt, 0])
+    for arr in calc_times.values():
+        arr[3] = monitor.task_no
+    dic = grid_point_sources(sources, params['ps_grid_spacing'], monitor)
+    for src in dic[grp_id]:
+        is_ps = isinstance(src, PointSource)
+        if is_ps:
+            src.nsites = srcfilter.sitecol.count_close(
+                src.location, md + pd) or EPS
+        else:
+            src.nsites = len(srcfilter.close_sids(src)) or EPS
+        src.num_ruptures = src.count_ruptures()
+        if pd and is_ps:
+            nphc = src.count_nphc()
+            if nphc > 1:
+                src.num_ruptures *= get_factor(
+                    nphc, src.location,
+                    srcfilter.sitecol, pd * 1.5, md + pd)
+    dic['calc_times'] = calc_times
+    dic['before'] = len(sources)
+    dic['after'] = len(dic[grp_id])
+    return dic
+
+
+def classical(srcs, rlzs_by_gsim, params, monitor):
+    """
+    Read the SourceFilter and call the classical calculator in hazardlib
+    """
+    srcfilter = monitor.read('srcfilter')
+    return hazclassical(srcs, srcfilter, rlzs_by_gsim, params, monitor)
 
 
 @base.calculators.add('classical', 'preclassical', 'ucerf_classical')
@@ -270,6 +377,7 @@ class ClassicalCalculator(base.HazardCalculator):
             self.datastore.swmr_on()  # fixes HDF5 error in build_hazard
             return
 
+        _weight_sources(self.csm, oq, self.datastore)
         smap = parallel.Starmap(classical, h5=self.datastore.hdf5)
         smap.monitor.save('srcfilter', self.src_filter())
         self.submit_tasks(smap)
